+++ CHANGELOG.md
@@ -21,7 +21,6 @@ The current status of the libraries at the time of the release is as follows:
 
 - Protocol buffers for hardware wallet transactions are no longer supported. Internet Computer Ledger app 2.4.9 or later is now required.
 - GovernanceCanister.listNeurons no longer throws an error when called with `certified: false` for hardware wallet transactions.
-- The field `smart_contract_address` was removed from ckETH minter information, and additional fields related to ERC20 support have been added.
 - Potential errors thrown by `withdrawEth` provide now unparsed details.
 
 ## Features
@@ -33,6 +32,7 @@ The current status of the libraries at the time of the release is as follows:
 - Remove dependency on `@dfinity/nns-proto` from `@dfinity/ledger-icp`.
 - Remove hardware wallet specific code and `@dfinity/nns-proto` dependency from `@dfinity/nns`.
 - Add support for `withdrawErc20`.
+- Additional fields related to ERC20 have been added to ckETH `minterInfo`.
 
 ## Build